# ketroy_back_new/storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "API документация Ketroy",
        "description": "Документация API для управления администраторами.",
        "version": "1.0"
    },
    "paths": {
...
        "/shops": {
            "get": {
                "tags": [
                    "Shops"
                ],
                "summary": "Получение списка магазинов",
                "description": "Возвращает список всех магазинов.",
                "operationId": "16f2967ba4e0c66824be1862e3ca3438",
                "responses": {
                    "200": {
                        "description": "Список магазинов",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "array",
                                    "items": {
                                        "properties": {
                                            "id": {
                                                "type": "integer"
                                            },
                                            "name": {
                                                "type": "string"
                                            },
                                            "description": {
                                                "type": "string"
                                            },
                                            "city": {
                                                "type": "string"
                                            },
                                            "two_gis_address": {
                                                "type": "string"
                                            },
                                            "address": {
                                                "type": "string"
                                            },
                                            "opening_hours": {
                                                "type": "string"
                                            }
                                        },
                                        "type": "object"
                                    }
                                }
                            }
                        }
                    }
                }
            },
            "post": {
                "tags": [
                    "Shops"
                ],
                "summary": "Создание магазина",
                "description": "Создает новый магазин.",
                "operationId": "bf1ecbd2dd07c06d915c0a5480a6c4d3",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "name",
                                    "city",
                                    "opening_hours"
                                ],
                                "properties": {
                                    "name": {
                                        "type": "string"
                                    },
                                    "description": {
                                        "type": "string"
                                    },
                                    "city": {
                                        "type": "string"
                                    },
                                    "two_gis_address": {
                                        "type": "string"
                                    },
                                    "address": {
                                        "type": "string"
                                    },
                                    "opening_hours": {
                                        "type": "string"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Магазин успешно создан",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Магазин успешно создан"
                                        },
                                        "shop": {
                                            "properties": {
                                                "id": {
                                                    "type": "integer"
                                                },
                                                "name": {
                                                    "type": "string"
                                                },
                                                "description": {
                                                    "type": "string"
                                                },
                                                "city": {
                                                    "type": "string"
                                                },
                                                "two_gis_address": {
                                                    "type": "string"
                                                },
                                                "address": {
                                                    "type": "string"
                                                },
                                                "opening_hours": {
                                                    "type": "string"
                                                }
                                            },
                                            "type": "object"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "400": {
                        "description": "Ошибка валидации"
                    }
                }
            }
        },
        "/shops/{id}": {
            "get": {
                "tags": [
                    "Shops"
                ],
                "summary": "Получение магазина",
                "description": "Возвращает магазин по указанному идентификатору.",
                "operationId": "722d75f836f7cfe561d0033a8a14986a",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID магазина",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Детали магазина",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "id": {
                                            "type": "integer"
                                        },
                                        "name": {
                                            "type": "string"
                                        },
                                        "description": {
                                            "type": "string"
                                        },
                                        "city": {
                                            "type": "string"
                                        },
                                        "two_gis_address": {
                                            "type": "string"
                                        },
                                        "address": {
                                            "type": "string"
                                        },
                                        "opening_hours": {
                                            "type": "string"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Магазин не найден"
                    }
                }
            },
            "put": {
                "tags": [
                    "Shops"
                ],
                "summary": "Обновление магазина",
                "description": "Обновляет существующий магазин.",
                "operationId": "5fdfbbcc6843c5b90dc7f7820dde3061",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID магазина",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "name",
                                    "city",
                                    "opening_hours"
                                ],
                                "properties": {
                                    "name": {
                                        "type": "string"
                                    },
                                    "description": {
                                        "type": "string"
                                    },
                                    "city": {
                                        "type": "string"
                                    },
                                    "two_gis_address": {
                                        "type": "string"
                                    },
                                    "address": {
                                        "type": "string"
                                    },
                                    "opening_hours": {
                                        "type": "string"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Магазин успешно обновлен",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Магазин успешно обновлён"
                                        },
                                        "shop": {
                                            "properties": {
                                                "id": {
                                                    "type": "integer"
                                                },
                                                "name": {
                                                    "type": "string"
                                                },
                                                "description": {
                                                    "type": "string"
                                                },
                                                "city": {
                                                    "type": "string"
                                                },
                                                "two_gis_address": {
                                                    "type": "string"
                                                },
                                                "address": {
                                                    "type": "string"
                                                },
                                                "opening_hours": {
                                                    "type": "string"
                                                }
                                            },
                                            "type": "object"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Магазин не найден"
                    },
                    "400": {
                        "description": "Ошибка валидации"
                    }
                }
            },
            "delete": {
                "tags": [
                    "Shops"
                ],
                "summary": "Удаление магазина",
                "description": "Удаляет магазин по указанному идентификатору.",
                "operationId": "708ed70ac0a9ab3e4b97025b4cf4e9f3",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID магазина",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "204": {
                        "description": "Магазин успешно удален"
                    },
                    "404": {
                        "description": "Магазин не найден"
                    }
                }
            }
        },
...
    },
...
}